+++ Packages/com.unity.inputsystem/InputSystem/Editor/AssetEditor/PropertiesViewBase.cs
@@ -4,6 +4,8 @@ using UnityEditor;
 using UnityEngine.Experimental.Input.Editor.Lists;
 using UnityEngine.Experimental.Input.Utilities;
 
+////TODO: show parameters for selected interaction or processor inline in list rather than separately underneath list
+
 namespace UnityEngine.Experimental.Input.Editor
 {
     /// <summary>
@@ -19,7 +21,7 @@ namespace UnityEngine.Experimental.Input.Editor
             m_InteractionsProperty = bindingOrAction.FindPropertyRelative("m_Interactions");
             m_ProcessorsProperty = bindingOrAction.FindPropertyRelative("m_Processors");
 
-            m_InteractionsList = new InteractionsListView(m_InteractionsProperty, OnInteractionsModified, expectedControlLayout);
+            m_InteractionsList = new InteractionsListView(m_InteractionsProperty, OnInteractionsModified, null);
             m_ProcessorsList = new ProcessorsListView(m_ProcessorsProperty, OnProcessorsModified, expectedControlLayout);
 
             m_OnChange = onChange;