+++ Packages/com.unity.inputsystem/Tests/InputSystem/CoreTests_Actions.cs
@@ -2584,6 +2584,33 @@ partial class CoreTests
         Assert.That(receivedVector.Value.y, Is.EqualTo(0.5678).Within(0.00001));
     }
 
+    [Test]
+    [Category("Actions")]
+    public void Actions_IncompatibleProcessorIsIgnored()
+    {
+        var gamepad = InputSystem.AddDevice<Gamepad>();
+
+        InputSystem.RegisterControlProcessor<ConstantVector2TestProcessor>();
+        var action = new InputAction(processors: "ConstantVector2Test");
+        action.AddBinding("<Gamepad>/leftStick/x");
+        action.Enable();
+
+        float? receivedFloat = null;
+        action.performed +=
+            ctx =>
+        {
+            Assert.That(receivedFloat, Is.Null);
+            // ConstantVector2TestProcessor processes Vector2s. It would throw an exception when
+            // trying to use it reading a float if not ignored.
+            receivedFloat = ctx.ReadValue<float>();
+        };
+
+        Set(gamepad.leftStick, Vector2.one);
+
+        Assert.That(receivedFloat, Is.Not.Null);
+        Assert.That(receivedFloat.Value, Is.EqualTo(1).Within(0.00001));
+    }
+
     // ReSharper disable once ClassNeverInstantiated.Local
     private class ConstantVector2TestProcessor : InputProcessor<Vector2>
     {
@@ -3579,13 +3606,11 @@ partial class CoreTests
         Assert.That(asset.FindAction($"{{{action3.id.ToString()}}}"), Is.SameAs(action3));
 
         // Shouldn't allocate.
-        #if UNITY_2018_3_OR_NEWER
         var map1action1 = "map1/action1";
         Assert.That(() =>
         {
             asset.FindAction(map1action1);
         }, Is.Not.AllocatingGCMemory());
-        #endif
     }
 
     [Test]
@@ -3932,7 +3957,6 @@ partial class CoreTests
         }
     }
 
-    #if UNITY_2018_3_OR_NEWER
     [Test]
     [Category("Actions")]
     [Ignore("TODO")]
@@ -3957,8 +3981,6 @@ partial class CoreTests
             }, Is.Not.AllocatingGCMemory());
         }
     }
-
-    #endif
 
     [Test]
     [Category("Actions")]